requirements.txt
apache-airflow-providers-amazon[s3fs]
apache-airflow-providers-google[gcs]
